Start to refactor into smaller methods

+++ lib/busboy/installer.rb
@@ -40,35 +40,48 @@ module Busboy
           File.read(manifest).split.each do |line|
 
             databag, item = line.split('/')
-            location = File.join("#{site[:value]}", 'databags', databag, item)
 
-            Busboy.logger.log "Downloading data bag item #{databag}/#{item} from '#{location}'"
+            databagpath = File.join(containerpath, databag)
+            Dir.mkdir(databagpath) unless Dir.exists?(databagpath)
 
-            response = Faraday.get(location)
-            if response.success?
-              begin
-                databagpath = File.join(containerpath, databag)
-                Dir.mkdir(databagpath) unless Dir.exists?(databagpath)
+            itempath = File.join(databagpath, "#{item}.json")
 
-                itempath = File.join(databagpath, "#{item}.json")
-                Busboy.logger.debug "Creating data bag item "#{itempath}"
+            unless File.exists?(itempath)
+              location = File.join("#{site[:value]}", 'databags', databag, item)
 
-                item = File.open(itempath, 'w')
-                item.write(response.body)
-                item.close
-              rescue SystemCallError, IOError => ex
-                raise Busboy::DatabagWriteError.new(databagpath), ex
+              Busboy.logger.log "Downloading data bag item #{databag}/#{item} from '#{location}'"
+
+              response = Faraday.get(location)
+              if response.success?
+                begin
+                  Busboy.logger.debug "Creating data bag item "#{itempath}"
+
+                  item = File.open(itempath, 'w')
+                  item.write(response.body)
+                  item.close
+                rescue SystemCallError, IOError => ex
+                  raise Busboy::DatabagWriteError.new(databagpath), ex
+                end
+              else
+                Busboy.logger.log "Failed to download #{location}"
+                # Raise an exception?
               end
+
             else
-              Busboy.logger.log "Failed to download #{location}"
-              # Raise an exception?
+              Busboy.logger.verbose "Skipping download of #{itempath} as it already exists"
             end
+
           end
+
         rescue SystemCallError, IOError => ex
           raise Busboy::DatabagReadError.new(manifest), ex
         end
 
       end
+    end
+
+    def download_item(site, databag, item)
+
     end
 
   end
